+++ hubdle-gradle-plugin/testFunctional/kotlin/com/javiersc/hubdle/project/config/VersioningTest.kt
@@ -6,9 +6,12 @@ import kotlin.test.Test
 internal class VersioningTest : GradleTestKitTest() {
 
     @Test
-    fun `given a project with versioning with semver enabled when it builds then SemverExtension exists`() {
+    fun `given a project with semver enabled when it builds then tag prefix is v`() {
         gradleTestKitTest("config/versioning/semver-1") {
-            gradlew("assemble").output.contains("semver.tagPrefix: v")
+            gradlew("assemble").output.apply {
+                contains("semver.tagPrefix: v")
+                contains("semver.version: 1.0.0")
+            }
         }
     }
 }